+++ src/Abstract.FileSystem/StringPathExtensions.cs
@@ -14,5 +14,25 @@
         {
             return SystemPath.FormatPath(path);
         }
+
+        /// <summary>
+        /// Remove the leading slash from the SystemPath
+        /// </summary>
+        /// <param name="path"></param>
+        /// <returns></returns>
+        public static string RemoveLeadingSlash(this SystemPath path)
+        {
+            return path.ToString().RemoveLeadingSlash();
+        }
+
+        /// <summary>
+        /// Remove the trailing slash from the SystemPath
+        /// </summary>
+        /// <param name="path"></param>
+        /// <returns></returns>
+        public static string RemoveTrailingSlash(this SystemPath path)
+        {
+            return path.ToString().RemoveTrailingSlash();
+        }
     }
 }