+++ src/Tests/Abstract.FileSystem.Test/SystemPathTests.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Abstract.FileSystem.Test
+﻿namespace Abstract.FileSystem.Test
 {
     public class SystemPathTests
     {